+++ CHANGES.rst
@@ -6,6 +6,10 @@ There's a frood who really knows where his towel is.
 1.5b2 (unreleased)
 ^^^^^^^^^^^^^^^^^^
 
+- Code clean up;
+  enforce usage of Products.PortalTransforms >=2.1.10 as ``<iframe>`` must be listed as a valid tag.
+  [hvelarde]
+
 - Remove hard dependency on collective.dexteritytextindexer.
   [hvelarde]